dailybot_core/requirements.txt
requests~=2.31.0
openai~=1.50.0
python-dotenv~=1.0.1
lark-oapi~=1.2.9
loguru~=0.7.2
PyYAML~=6.0.1
httpx~=0.27.0
pydantic>=2.0.0
